requirements.txt
Flask
Pillow
requests